Make sure Renderer destructor doesnt throw
Renderer destructor could throw exception before and so we had to seperate the waiting for fences to another seperate function. We also made sure to check if the startup vulkan objects are null or not before attempting to destroy them.

+++ src/VulkanFoundationalElements.cpp
@@ -9,8 +9,14 @@ namespace VRenderer
 {
 	void VulkanFoundationalElements::CleanUp()
 	{
-		vkDestroySurfaceKHR(m_instance, m_surface, nullptr);
-		vkb::destroy_debug_utils_messenger(m_instance, m_debugMsger);
-		vkDestroyInstance(m_instance, nullptr);
+		if (VK_NULL_HANDLE != m_instance) {
+			if (VK_NULL_HANDLE != m_surface) {
+				vkDestroySurfaceKHR(m_instance, m_surface, nullptr);
+			}
+			if (VK_NULL_HANDLE != m_debugMsger) {
+				vkb::destroy_debug_utils_messenger(m_instance, m_debugMsger);
+			}
+			vkDestroyInstance(m_instance, nullptr);
+		}
 	}
 }
